test+fix: add unit tests, CI, and fix three production bugs found via audit
Tests (40 total, all passing):
- i18n parity gate: vi and en locales must have identical keys
- parse helpers (k/m/b suffix, comma separators, fmt)
- symbol registry + commodity metadata
- alert matching logic (above/below, boundary cases)
- pnl computation for long/short
- DB schema smoke test against SQLite in-memory
- payment webhook idempotency (duplicate tx_id is no-op)
- subscription lifecycle: fresh activate, renewal stacks from existing
  expiry, creates missing user, expired premium renews from now

Bug fixes:
1. Payment idempotency: Telegram can retry successful_payment webhooks.
   Previously, activate_premium ran unconditionally after on_conflict_do_nothing,
   so a retry would double-extend the user's premium window. Now we gate
   activation on result.rowcount==1 from the upsert. Covered by test.
2. Integer PK for Alert/Trade/Payment: BigInteger primary keys are not
   auto-incremented on SQLite (only INTEGER PRIMARY KEY is a ROWID alias),
   so local dev on SQLite would fail on first insert. Switched surrogate
   ids to Integer while keeping BigInteger for telegram user_id FKs.
3. SQLite timezone handling: SQLite strips tzinfo on read, causing
   "can't subtract naive and aware datetimes" in is_premium and
   expire_and_remind. Added _aware() normalizer in subscription service
   and defensive upgrade in users.is_premium.

CI: .github/workflows/ci.yml runs ruff + i18n parity gate + pytest on
every push and PR against main.

+++ bot/models.py
@@ -1,7 +1,7 @@
 from datetime import datetime
 from decimal import Decimal
 
-from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Numeric, String, Text, func
+from sqlalchemy import BigInteger, Boolean, DateTime, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
 from bot.db import Base
@@ -27,7 +27,7 @@ class User(Base):
 class Alert(Base):
     __tablename__ = "alerts"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.tg_id", ondelete="CASCADE"), index=True)
     asset_class: Mapped[str] = mapped_column(String(16), nullable=False)  # crypto|metal|energy
     symbol: Mapped[str] = mapped_column(String(32), nullable=False, index=True)
@@ -46,7 +46,7 @@ class Alert(Base):
 class Trade(Base):
     __tablename__ = "trades"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.tg_id", ondelete="CASCADE"), index=True)
     asset_class: Mapped[str] = mapped_column(String(16), nullable=False)
     symbol: Mapped[str] = mapped_column(String(32), nullable=False)
@@ -70,7 +70,7 @@ class Trade(Base):
 class Payment(Base):
     __tablename__ = "payments"
 
-    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
     user_id: Mapped[int] = mapped_column(ForeignKey("users.tg_id", ondelete="CASCADE"), index=True)
     provider: Mapped[str] = mapped_column(String(32), nullable=False)
     amount: Mapped[Decimal] = mapped_column(Numeric(20, 8), nullable=False)
